[MOD] Admin prefix into routes

+++ routes/web.php
@@ -12,15 +12,14 @@ use Illuminate\Support\Facades\Route;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
-
-//TODO Rename to add 'admin' prefix to path
-Route::get('/items/create', 'App\Http\Controllers\ItemController@create')->name('item.create');
-Route::post('/items/save', 'App\Http\Controllers\ItemController@save')->name('item.save');
-Route::delete('/items/delete/{id}', 'App\Http\Controllers\ItemController@delete')->name('item.delete');
-Route::get('/items', 'App\Http\Controllers\ItemController@index')->name('item.index');
-Route::get('/items/{id}', 'App\Http\Controllers\ItemController@show')->name('item.show');
-
 Route::get('/admin', 'App\Http\Controllers\AdminController@index')->name('admin.index');
+
+Route::get('/admin/items/create', 'App\Http\Controllers\ItemController@create')->name('admin.item.create');
+Route::post('/admin/items/save', 'App\Http\Controllers\ItemController@save')->name('admin.item.save');
+Route::delete('/admin/items/delete/{id}', 'App\Http\Controllers\ItemController@delete')->name('admin.item.delete');
+Route::get('/admin/items', 'App\Http\Controllers\ItemController@index')->name('admin.item.index');
+Route::get('/admin/items/{id}', 'App\Http\Controllers\ItemController@show')->name('admin.item.show');
+
 Route::get('/admin/reviews', 'App\Http\Controllers\ReviewController@index')->name('admin.review.index');
 Route::get('/admin/reviews/create', 'App\Http\Controllers\ReviewController@create')->name('admin.review.create');
 Route::post('/admin/reviews/save', 'App\Http\Controllers\ReviewController@save')->name('admin.review.save');
